+++ src/Contracts/AcceptsTeamInvitations.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Filament\Jetstream\Contracts;
+
+use Illuminate\Http\RedirectResponse;
+
+/**
+ * @method RedirectResponse accept(string|int $invitationId)
+ */
+interface AcceptsTeamInvitations
+{
+    //
+}